+++ CotEditor/Sources/Constants.h
@@ -240,6 +240,7 @@ extern NSString *__nonnull const CEDefaultColoringRangeBufferLengthKey;
 extern NSString *__nonnull const CEDefaultLargeFileAlertThresholdKey;
 extern NSString *__nonnull const CEDefaultAutosavingDelayKey;
 extern NSString *__nonnull const CEDefaultEnablesAutosaveInPlaceKey;
+extern NSString *__nonnull const CEDefaultSavesTextOrientationKey;