+++ CHANGELOG.md
@@ -10,6 +10,8 @@ develop
 - CotEditor is now __Sandboxed__.
 - New setting option for the behavior on document modification by external process (in Genearal pane).
 - Share button in toolbar (Customize toobar to use it).
+- Save text orientation state to the file and restore it when the file is opened.
+    - __for advanced users__: In this feature, CotEditor saves an *extended attribute* which named `com.coteditor.VerticalText` to the file only when the editor's text orientation is vertical. You can even disable the feature running the command `defaults write com.coteditor.CotEditor savesTextOrientation -bool NO` in Terminal. 
 
 
 ### Additions/Changes